done:  - rake assemble  - rake watch  - rake clean  - rake package  - rake tag  - Jenkinsfile
ready to test :D

+++ scripts/clean.sh
@@ -12,6 +12,9 @@ rm -rf www/*
 # remove all old node modules
 rm -rf node_modules/
 
-# restore original project config file
-rm -rf ./ionic.config.json
-mv ./ionic.config.old.json ./ionic.config.json
+# check if original project config was modified
+if [ -f ./ionic.config.old.json ]; then
+  # restore original project config file
+  rm -rf ./ionic.config.json
+  mv ./ionic.config.old.json ./ionic.config.json
+fi
